requirements.txt
dagster
dagster-webserver
pandas
requests
python-dotenv
pytest
matplotlib
sqlalchemy
psycopg2-binary